+++ config.py
@@ -130,6 +130,7 @@ class Config:
     CACHE_REDIS_HOST = config.get("REDIS_HOST", "localhost")
     CACHE_REDIS_PORT = config.get("REDIS_PORT", 6379)
     CACHE_REDIS_DB = config.get("REDIS_DB", 0)
+    DEFAULT_TIMEZONE = config.get("DEFAULT_TIMEZONE", "Europe/Helsinki")
     
     @classmethod
     def init_config(cls) -> None: